move most metrics-related infrastructure to the "core" module
