+++ implementation/core/src/main/java/io/smallrye/faulttolerance/core/metrics/MeteredOperation.java
@@ -1,0 +1,19 @@
+package io.smallrye.faulttolerance.core.metrics;
+
+public interface MeteredOperation {
+    boolean isAsynchronous();
+
+    boolean hasBulkhead();
+
+    boolean hasCircuitBreaker();
+
+    boolean hasFallback();
+
+    boolean hasRateLimit();
+
+    boolean hasRetry();
+
+    boolean hasTimeout();
+
+    String name();
+}
