+++ implementation/core/src/main/java/io/smallrye/faulttolerance/core/metrics/MetricsProvider.java
@@ -1,0 +1,7 @@
+package io.smallrye.faulttolerance.core.metrics;
+
+public interface MetricsProvider {
+    boolean isEnabled();
+
+    MetricsRecorder create(MeteredOperation operation);
+}
