+++ implementation/fault-tolerance/src/main/java/io/smallrye/faulttolerance/metrics/NoopProvider.java
@@ -2,7 +2,8 @@ package io.smallrye.faulttolerance.metrics;
 
 import jakarta.inject.Singleton;
 
-import io.smallrye.faulttolerance.config.FaultToleranceOperation;
+import io.smallrye.faulttolerance.core.metrics.MeteredOperation;
+import io.smallrye.faulttolerance.core.metrics.MetricsProvider;
 import io.smallrye.faulttolerance.core.metrics.MetricsRecorder;
 
 @Singleton
@@ -13,7 +14,7 @@ public class NoopProvider implements MetricsProvider {
     }
 
     @Override
-    public MetricsRecorder create(FaultToleranceOperation operation) {
+    public MetricsRecorder create(MeteredOperation operation) {
         return MetricsRecorder.NOOP;
     }
 }
